[#27340] Add RestUserController (/register, /findByEmail  and /findAll  -> OK)   and  change int to double in value of CurrencyHistorical

+++ src/main/java/com/capitalmarkets/app/integration/adapters/Impl/CurrencyAdapter.java
@@ -46,7 +46,7 @@ public class CurrencyAdapter implements ICurrencyAdapter {
     }
 
     @Override
-    public CurrencyHistoricalDTO getHistorical(String date, int value, String base, String conversion) {
+    public CurrencyHistoricalDTO getHistorical(String date, double value, String base, String conversion) {
 
         restTemplate.getForObject(url+date+"..?amount="+value+"&from="+base+"&to="+conversion,String.class);
         log.info(restTemplate.getForObject(url+date+"..?amount="+value+"&from="+base+"&to="+conversion,String.class));
